+++ bindings/nostr-ffi/src/nips/nip01.rs
@@ -1,6 +1,7 @@
 // Copyright (c) 2023-2024 Rust Nostr Developers
 // Distributed under the MIT software license
 
+use std::fmt;
 use std::ops::Deref;
 use std::str::FromStr;
 use std::sync::Arc;
@@ -15,9 +16,15 @@ use crate::{Kind, PublicKey};
 
 /// Coordinate for event (`a` tag)
 #[derive(Debug, PartialEq, Eq, Hash, Object)]
-#[uniffi::export(Debug, Eq, Hash)]
+#[uniffi::export(Debug, Display, Eq, Hash)]
 pub struct Coordinate {
     inner: nip01::Coordinate,
+}
+
+impl fmt::Display for Coordinate {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.inner)
+    }
 }
 
 impl Deref for Coordinate {