+++ bindings/nostr-ffi/tests/test_generated_bindings.rs
@@ -1,1 +1,5 @@
-uniffi::build_foreign_language_testcases!("tests/test_equality.py", "tests/test_nip05.py");
+uniffi::build_foreign_language_testcases!(
+    "tests/test_equality.py",
+    "tests/test_nip05.py",
+    "tests/test_to_string.py",
+);